v0.4.0
changelog:
- combined RandomForest and XGBoost into one single module namely, tree
