+++ logs/cov.out
@@ -1,11 +1,10 @@
-Name                                                             Stmts   Miss  Cover   Missing
-----------------------------------------------------------------------------------------------
-/media/ph33r/Data/Project/mllib/Git/mllib/__init__.py                7      0   100%
-/media/ph33r/Data/Project/mllib/Git/mllib/lib/__init__.py            7      0   100%
-/media/ph33r/Data/Project/mllib/Git/mllib/lib/boost.py              53      0   100%
-/media/ph33r/Data/Project/mllib/Git/mllib/lib/cluster.py           103      0   100%
-/media/ph33r/Data/Project/mllib/Git/mllib/lib/knn.py                70      0   100%
-/media/ph33r/Data/Project/mllib/Git/mllib/lib/model.py              44      0   100%
-/media/ph33r/Data/Project/mllib/Git/mllib/lib/random_forest.py      61      0   100%
-----------------------------------------------------------------------------------------------
-TOTAL                                                              345      0   100%
+Name                                                        Stmts   Miss  Cover   Missing
+-----------------------------------------------------------------------------------------
+/media/ph33r/Data/Project/mllib/Git/mllib/__init__.py           7      0   100%
+/media/ph33r/Data/Project/mllib/Git/mllib/lib/__init__.py       7      0   100%
+/media/ph33r/Data/Project/mllib/Git/mllib/lib/cluster.py      103      0   100%
+/media/ph33r/Data/Project/mllib/Git/mllib/lib/knn.py           70      0   100%
+/media/ph33r/Data/Project/mllib/Git/mllib/lib/model.py         44      0   100%
+/media/ph33r/Data/Project/mllib/Git/mllib/lib/tree.py          79      0   100%
+-----------------------------------------------------------------------------------------
+TOTAL                                                         310      0   100%
