--- logs/pylint/tests-test_random_forest-py.out
@@ -1,4 +1,0 @@
-
---------------------------------------------------------------------
-Your code has been rated at 10.00/10 (previous run: 10.00/10, +0.00)
-
