+++ requirements.txt
@@ -1,5 +1,5 @@
-pandas==1.1.3
 Cython==0.29.15
-xgboost==1.3.3
 numpy==1.19.5
+pandas==1.1.3
+xgboost==1.3.3
 scikit_learn==1.0
