+++ tests/test_tree.py
@@ -34,7 +34,8 @@ path = re.sub(r"(.+)(\/tests.*)", "\\1", path)
 
 sys.path.insert(0, path)
 
-from mllib.lib.boost import XGBoost  # noqa: F841
+from mllib.lib.tree import RandomForest  # noqa: F841
+from mllib.lib.tree import XGBoost  # noqa: F841
 
 # =============================================================================
 # --- DO NOT CHANGE ANYTHING FROM HERE
@@ -57,6 +58,46 @@ def ignore_warnings(test_func):
     return do_test
 
 
+class Test_RandomForest(unittest.TestCase):
+    """Test suite for module ``RandomForest``."""
+
+    def setUp(self):
+        """Set up for module ``RandomForest``."""
+
+    def test_rf_class(self):
+        """RandomForest: Test for classification."""
+        x_var = ["x1", "x2", "x3", "x4"]
+        y_var = "y"
+        df_ip = pd.read_csv(path + "iris.csv")
+        df_ip = df_ip[[y_var] + x_var]
+        df_train, df_test = split(df_ip,
+                                  stratify=df_ip[y_var],
+                                  test_size=0.2,
+                                  random_state=42)
+        mod = RandomForest(df_train, y_var, x_var, method="classify")
+        y_hat = mod.predict(df_test[x_var])[y_var].tolist()
+        y = df_test[y_var].values.tolist()
+        acc = round(sk_metrics.accuracy_score(y, y_hat), 2)
+        self.assertGreaterEqual(acc, 0.93)
+
+    @ignore_warnings
+    def test_rf_reg(self):
+        """RandomForest: Test for regression."""
+        x_var = ["x1", "x2", "x3", "x4"]
+        y_var = "y"
+        df_ip = pd.read_csv(path + "iris.csv")
+        df_ip = df_ip[[y_var] + x_var]
+        df_train, df_test = split(df_ip,
+                                  stratify=df_ip[y_var],
+                                  test_size=0.2,
+                                  random_state=42)
+        mod = RandomForest(df_train, y_var, x_var, method="regression")
+        y_hat = mod.predict(df_test[x_var])[y_var].tolist()
+        y = df_test[y_var].values.tolist()
+        mse = round(sk_metrics.mean_squared_error(y, y_hat), 2)
+        self.assertLessEqual(mse, 0.1)
+
+
 class Test_XGBoost(unittest.TestCase):
     """Test suite for module ``XGBoost``."""
 
@@ -66,7 +107,7 @@ class Test_XGBoost(unittest.TestCase):
     @ignore_warnings
     def test_xgboost_class(self):
         """XGBoost: Test for classification."""
-        x_var = ["x1", "x2", "x3", "x4"]
+        x_var = ["x1", "x2"]
         y_var = "y"
         df_ip = pd.read_csv(path + "iris.csv")
         df_ip = df_ip[[y_var] + x_var]
@@ -74,9 +115,7 @@ class Test_XGBoost(unittest.TestCase):
                                   stratify=df_ip[y_var],
                                   test_size=0.2,
                                   random_state=1)
-        mod = XGBoost(df_train, y_var, x_var, method="classify",
-                      param={"n_estimators": [1],
-                             "objective": ["binary:logistic"]})
+        mod = XGBoost(df_train, y_var, x_var, method="classify")
         y_hat = mod.predict(df_test[x_var])[y_var].tolist()
         y = df_test[y_var].values.tolist()
         acc = round(sk_metrics.accuracy_score(y, y_hat), 2)
@@ -93,9 +132,7 @@ class Test_XGBoost(unittest.TestCase):
                                   stratify=df_ip[y_var],
                                   test_size=0.2,
                                   random_state=1)
-        mod = XGBoost(df_train, y_var, x_var, method="regression",
-                      param={"n_estimators": [1],
-                             "objective": ["reg:squarederror"]})
+        mod = XGBoost(df_train, y_var, x_var, method="regression")
         y_hat = mod.predict(df_test[x_var])[y_var].tolist()
         y = df_test[y_var].values.tolist()
         mse = round(sk_metrics.mean_squared_error(y, y_hat), 2)
